Logger context

+++ SmartHouse.Domain.Core/WeatherModel.cs
@@ -2,21 +2,21 @@
 {
     public class WeatherModel
     {
-        public int WindSpeed { get; set; }
-        public int WindDeg { get; set; }
-        public double Temp { get; set; }
+        public float WindSpeed { get; set; }
+        public ushort WindDeg { get; set; }
+        public float Temp { get; set; }
         public string City { get; set; }
-        public int Pressure { get; set; }
-        public int Humidity { get; set; }
+        public float Pressure { get; set; }
+        public float Humidity { get; set; }
         public string Description { get; set; }
-        public double FeelsLike { get; set; }
+        public float FeelsLike { get; set; }
 
         public WeatherModel()
         {
           
         }
 
-        public WeatherModel(int windSpeed, int windDeg, double temp, string city, int pressure, int humidity, string description, double feelsLike)
+        public WeatherModel(float windSpeed, byte windDeg, float temp, string city, float pressure, float humidity, string description, float feelsLike)
         {
             WindSpeed = windSpeed;
             WindDeg = windDeg;
