+++ SmartHouse.Service.Weather.OpenWeatherService/Model/WeatherMain.cs
@@ -5,16 +5,16 @@ namespace SmartHouse.Service.Weather.OpenWeatherMap.Model
     public class Main
     {
         [JsonPropertyName("temp")]
-        public double Temp { get; set; }
+        public float Temp { get; set; }
         [JsonPropertyName("feels_like")]
-        public double FeelsLike { get; set; }
+        public float FeelsLike { get; set; }
         [JsonPropertyName("temp_min")]
-        public double TempMin { get; set; }
+        public float TempMin { get; set; }
         [JsonPropertyName("temp_max")]
-        public double TempMax { get; set; }
+        public float TempMax { get; set; }
         [JsonPropertyName("pressure")]
-        public int Pressure { get; set; }
+        public float Pressure { get; set; }
         [JsonPropertyName("humidity")]
-        public int Humidity { get; set; }
+        public float Humidity { get; set; }
     }
 }
