+++ SmartHouse.Service.Weather.OpenWeatherService/Model/Wind.cs
@@ -5,8 +5,8 @@ namespace SmartHouse.Service.Weather.OpenWeatherMap.Model
     public class Wind
     {
         [JsonPropertyName("speed")]
-        public double Speed { get; set; }
+        public float Speed { get; set; }
         [JsonPropertyName("deg")]
-        public int Deg { get; set; }
+        public ushort Deg { get; set; }
     }
 }
